move distributed/proxynode to distributed/proxy
