+++ internal/distributed/rootcoord/service.go
@@ -26,7 +26,7 @@ import (
 	grpc_opentracing "github.com/grpc-ecosystem/go-grpc-middleware/tracing/opentracing"
 	dsc "github.com/milvus-io/milvus/internal/distributed/datacoord/client"
 	isc "github.com/milvus-io/milvus/internal/distributed/indexcoord/client"
-	pnc "github.com/milvus-io/milvus/internal/distributed/proxynode/client"
+	pnc "github.com/milvus-io/milvus/internal/distributed/proxy/client"
 	qsc "github.com/milvus-io/milvus/internal/distributed/queryservice/client"
 	"github.com/milvus-io/milvus/internal/log"
 	"github.com/milvus-io/milvus/internal/msgstream"
